COMU-3250 - Adding or editing register data in UI does not take overflow into account

+++ comu/com.energyict.mdc.device.data/src/main/java/com/energyict/mdc/device/data/impl/DeviceImpl.java
@@ -299,7 +299,7 @@ public class DeviceImpl implements Device, ServerDeviceForConfigChange, ServerDe
             Meter.MeterReadingTypeConfigurationBuilder meterReadingTypeConfigurationBuilder = meterConfigurationBuilder
                     .configureReadingType(channelSpec.getReadingType())
                     .withNumberOfFractionDigits(channelSpec.getNbrOfFractionDigits())
-                    .withOverflowValue(channelSpec.getOverflow().longValue());
+                    .withOverflowValue(channelSpec.getOverflow());
             if (channelSpec.isUseMultiplier()) {
                 meterReadingTypeConfigurationBuilder
                         .withMultiplierOfType(defaultMultiplierType)
@@ -314,7 +314,7 @@ public class DeviceImpl implements Device, ServerDeviceForConfigChange, ServerDe
             Meter.MeterReadingTypeConfigurationBuilder meterReadingTypeConfigurationBuilder = meterConfigurationBuilder
                     .configureReadingType(registerSpec.getReadingType())
                     .withNumberOfFractionDigits(registerSpec.getNumberOfFractionDigits())
-                    .withOverflowValue(registerSpec.getOverflowValue().longValue());
+                    .withOverflowValue(registerSpec.getOverflowValue());
             if (registerSpec.isUseMultiplier()) {
                 meterReadingTypeConfigurationBuilder
                         .withMultiplierOfType(defaultMultiplierType)
